Move `lambda/shared` to `internal/shared`
To make clear that it isn't a lambda itself, and to align more closely with other OPG Go repos' structure.

#patch
