+++ internal/shared/client.go
