+++ internal/shared/event.go
