+++ internal/shared/lpa.go
