+++ lambda/update/main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/go-openapi/jsonpointer"
-	"github.com/ministryofjustice/opg-data-lpa-store/lambda/shared"
+	"github.com/ministryofjustice/opg-data-lpa-store/internal/shared"
 	"github.com/ministryofjustice/opg-go-common/logging"
 )
